+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Make Errors::Builder thread-safe.
 
 ## [1.4.0] - 2016-08-15
 ### Added